# Setting.Service.UnchaseClient/Connected Services/OpenAPIService/OpenAPIService.nswag.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Setting.Service",
    "description": "API для работы с сервисом настроек",
    "termsOfService": "https://confluence.nsc-tech.ru/pages/viewpage.action?pageId=81952797",
    "contact": {
      "name": "PROC-4092",
      "url": "https://jira.nsc-tech.ru/browse/PROC-4108",
      "email": ""
    },
    "version": "1.0.0"
  },
  "paths": {
    "/Cache/Reset": {
      "get": {
        "tags": [
          "Cache"
        ],
        "summary": "Сброс кеша сервиса.",
        "description": "Пример использования:\r\n            \r\n    GET /cache/reset",
        "operationId": "Reset",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Неправильно сформирован запрос: не хватает обязательных параметров, используются неподдерживаемые параметры, отсутствует подпись или возникла другая ошибка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Маркер доступа (access token) просрочен, аннулирован или не действует по другим причинам.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Маркер доступа должен иметь более широкие границы применения. Ответ может содержать требуемое значение атрибута “scope”",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка входных данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка"
          }
        }
      }
    },
    "/Modules": {
      "get": {
        "tags": [
          "Modules"
        ],
        "summary": "Получение информации о модулях системы.",
        "description": "Пример использования:\r\n            \r\n    GET /modules",
        "operationId": "GetModules",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Module"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Неправильно сформирован запрос: не хватает обязательных параметров, используются неподдерживаемые параметры, отсутствует подпись или возникла другая ошибка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Маркер доступа (access token) просрочен, аннулирован или не действует по другим причинам.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Маркер доступа должен иметь более широкие границы применения. Ответ может содержать требуемое значение атрибута “scope”",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка входных данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка"
          }
        }
      }
    },
    "/Modules/{moduleId}/settings/{settingName}/value": {
      "get": {
        "tags": [
          "Modules"
        ],
        "summary": "Получение текстового значения настройки по названию и идентификатору модуля.",
        "description": "Пример использования:\r\n            \r\n    GET /modules/127/settings/DefaultCacheLifeTimeInMinutes/value",
        "operationId": "GetSettingValueByModuleId",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "description": "Идентификатор модуля",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "settingName",
            "in": "path",
            "required": true,
            "description": "Название настройки",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Неправильно сформирован запрос: не хватает обязательных параметров, используются неподдерживаемые параметры, отсутствует подпись или возникла другая ошибка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Маркер доступа (access token) просрочен, аннулирован или не действует по другим причинам.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Маркер доступа должен иметь более широкие границы применения. Ответ может содержать требуемое значение атрибута “scope”",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка входных данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка"
          }
        }
      }
    },
    "/Modules/{moduleId}/{useCache}/settings/values": {
      "get": {
        "tags": [
          "Modules"
        ],
        "summary": "Получение словаря текстовых значений настроек модуля по его идентификатору.",
        "description": "Пример использования:\r\n            \r\n    GET /modules/5/true/settings/values",
        "operationId": "GetSettingValuesByModuleId",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "description": "Идентификатор модуля",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "useCache",
            "in": "path",
            "required": true,
            "description": "Признак использования кеширования. Если false - получить значения из БД MasterData с обновлением кеша сервиса.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Setting"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Неправильно сформирован запрос: не хватает обязательных параметров, используются неподдерживаемые параметры, отсутствует подпись или возникла другая ошибка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Маркер доступа (access token) просрочен, аннулирован или не действует по другим причинам.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Маркер доступа должен иметь более широкие границы применения. Ответ может содержать требуемое значение атрибута “scope”",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка входных данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Type": {
            "type": "string",
            "nullable": true
          },
          "Title": {
            "type": "string",
            "nullable": true
          },
          "Status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "Detail": {
            "type": "string",
            "nullable": true
          },
          "Instance": {
            "type": "string",
            "nullable": true
          },
          "Extensions": {
            "type": "object",
            "readOnly": true,
            "nullable": true,
            "additionalProperties": {
              "type": "object",
              "additionalProperties": false
            }
          }
        }
      },
      "Module": {
        "type": "object",
        "description": "Модуль",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор модуля",
            "format": "int32"
          },
          "systemName": {
            "type": "string",
            "description": "Системное название модуля",
            "nullable": true
          }
        }
      },
      "Setting": {
        "type": "object",
        "description": "Настройка",
        "additionalProperties": false,
        "required": [
          "name",
          "value"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Название"
          },
          "value": {
            "type": "string",
            "description": "Значение"
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "description": "Для доступа к API необходимо авторизоваться!",
        "scheme": "basic"
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}